Primeblocks and non-convex constellation work
Two notebooks added.  Primeblocks contains code for quickly identifying arrays of consecutive primes among modestly large integers (say up to E19).  We use these arrays in another notebook (coming soon) to compare actual statistics for prime gaps to the estimates provided by the relative population models.

Engelsma contains an initial study of the images of a non-convex constellation across the cycles of gaps.  We generate the primorial coordinates for the locations of images across several stages of the sieve, and we calculate its asymptotic relative population. (Higher than we expected!)